requirements.txt
pandas>=1.3.0
numpy>=1.20.0
finnhub-python>=2.4.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0 